blogon: server-side app context, display unpublished posts in dev, adjust paths
The store is now setup in `main` on the server side, and knows about
whether we are in prod or not. This lets us display unpublished posts in
dev.

The path to the different assets, and the server_fn API are properly
namespaced.

As for the previous commit saying that the WASM bundle size is high,
this was wrong: it's pretty much the baseline of the framework
(uncompressed). You could go lower by using islands, but that's a
different rendering model.